+++ src/prelude.rs
@@ -5,7 +5,5 @@
 //! ```
 
 pub use crate::error::{Error, Result};
-#[cfg(feature = "stream")]
-pub use crate::service::Streaming;
-pub use crate::service::{Files, Readiness, Tasks};
-pub use crate::{Client, ClientBuilder, Endpoint};
+pub use crate::multipart::{Multipart, Part};
+pub use crate::{Client, ClientBuilder, Endpoint, EndpointResponse};